# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Posts API",
        "version": "0.1"
    },
    "paths": {
        "/posts": {
            "get": {
                "description": "Get a list of posts",
                "responses": {
                    "200": {
                        "description": "Get posts successfully !"
                    }
                }
            },
            "post": {
                "description": "Create a new posts",
                "requestBody": {
                    "description": "Data of the new posts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "type": "number"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "created_by": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create posts successfully !"
                    },
                    "422": {
                        "description": "There was an error during creation !"
                    }
                }
            }
        },
        "/posts/{$id}": {
            "put": {
                "description": "Update a posts",
                "requestBody": {
                    "description": "Data of the new posts update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "type": "number"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "updated_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_by": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update posts successfully !"
                    },
                    "422": {
                        "description": "There was an error during the update process !"
                    }
                }
            }
        }
    }
}